projects/capstone/starter/requirements.txt
alembic==1.3.3
Click==7.0
ecdsa==0.15
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.0
Mako==1.1.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycrypto==2.6.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==1.0.0
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.1
